decrept secret data for topology
