+++ controllers/servers/host_definer/kubernetes_manager/manager.py
@@ -1,4 +1,5 @@
 import datetime
+import base64
 
 from kubernetes import client, config, dynamic
 from kubernetes.client import api_client
@@ -286,13 +287,37 @@ class KubernetesManager():
 
     def _get_secret_data(self, secret_name, secret_namespace):
         try:
-            return self.core_api.read_namespaced_secret(name=secret_name, namespace=secret_namespace).data
+            secret_data = self.core_api.read_namespaced_secret(name=secret_name, namespace=secret_namespace).data
+            return self._change_decode_base64_secret_config(secret_data)
         except ApiException as ex:
             if ex.status == 404:
                 logger.error(messages.SECRET_DOES_NOT_EXIST.format(secret_name, secret_namespace))
             else:
                 logger.error(messages.FAILED_TO_GET_SECRET.format(secret_name, secret_namespace, ex.body))
-            return {}
+        except TypeError as ex:
+            logger.error(str(ex))
+        return {}
+
+    def _change_decode_base64_secret_config(self, secret_data):
+        if settings.SECRET_CONFIG_FIELD in secret_data.keys():
+            secret_data[settings.SECRET_CONFIG_FIELD] = self._decode_base64_to_dict(
+                secret_data[settings.SECRET_CONFIG_FIELD])
+        return secret_data
+
+    def _decode_base64_to_dict(self, content_with_base64):
+        decoded_string_content = self._decode_base64_to_string(content_with_base64)
+        encoded_dict = str(decoded_string_content).encode('utf-8')
+        base64_dict = base64.b64encode(encoded_dict)
+        my_dict_again = eval(base64.b64decode(base64_dict))
+        return my_dict_again
+
+    def _decode_base64_to_string(self, content_with_base64):
+        try:
+            base64_bytes = content_with_base64.encode('ascii')
+            decoded_string_in_bytes = base64.b64decode(base64_bytes)
+            return decoded_string_in_bytes.decode('ascii')
+        except Exception:
+            return content_with_base64
 
     def _read_node(self, node_name):
         try:
